fix: Property 'TextDecoder' doesn't exist
